pyproject.toml
[project]
name = "rundown"
version = "1.0.0"
description = "Generate beautiful weekly Strava running stats images"
dependencies = [
    "stravalib",
    "python-dotenv",
    "pillow",
    "python-dateutil"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
rundown = "rundown.main:main"